Move `ExternalDeploymentTaskTrigger` to core
Related to https://github.com/astronomer/astronomer-providers/pull/683#discussion_r983774615

This should be in the external_task.py
