+++ astronomer/providers/core/triggers/external_task.py
@@ -1,14 +1,17 @@
 import asyncio
 import datetime
 import typing
-from typing import Any, Dict, List, Tuple
+from typing import Any, AsyncIterator, Dict, List, Tuple
 
+from airflow import AirflowException
 from airflow.models import DagRun, TaskInstance
 from airflow.triggers.base import BaseTrigger, TriggerEvent
 from airflow.utils.session import provide_session
 from asgiref.sync import sync_to_async
 from sqlalchemy import func
 from sqlalchemy.orm import Session
+
+from astronomer.providers.http.triggers.http import HttpTrigger
 
 
 class TaskStateTrigger(BaseTrigger):
@@ -145,3 +148,47 @@ class DagStateTrigger(BaseTrigger):
             .scalar()
         )
         return typing.cast(int, count)
+
+
+class ExternalDeploymentTaskTrigger(HttpTrigger):
+    """ExternalDeploymentTaskTrigger Inherits from HttpTrigger and make Async http call to get the deployment state"""
+
+    def serialize(self) -> Tuple[str, Dict[str, Any]]:
+        """Serializes ExternalDeploymentTaskTrigger arguments and classpath."""
+        return (
+            "astronomer.providers.http.triggers.http.ExternalDeploymentTaskTrigger",
+            {
+                "endpoint": self.endpoint,
+                "data": self.data,
+                "headers": self.headers,
+                "extra_options": self.extra_options,
+                "http_conn_id": self.http_conn_id,
+                "poke_interval": self.poke_interval,
+            },
+        )
+
+    async def run(self) -> AsyncIterator["TriggerEvent"]:  # type: ignore[override]
+        """
+        Makes a series of asynchronous http calls via an http hook poll for state of the job
+        run until it reaches a failure state or success state. It yields a Trigger if response state is successful.
+        """
+        from airflow.utils.state import State
+
+        hook = self._get_async_hook()
+        while True:
+            try:
+                response = await hook.run(
+                    endpoint=self.endpoint,
+                    data=self.data,
+                    headers=self.headers,
+                    extra_options=self.extra_options,
+                )
+                resp_json = await response.json()
+                if resp_json["state"] in State.finished:
+                    yield TriggerEvent(resp_json)
+                else:
+                    await asyncio.sleep(self.poke_interval)
+            except AirflowException as exc:
+                if str(exc).startswith("404"):
+                    await asyncio.sleep(self.poke_interval)
+                yield TriggerEvent({"state": "error", "message": str(exc)})
