+++ docs/providers/core/sensors/external_task.rst
@@ -10,3 +10,16 @@ To waits asynchronously for a external task to complete
     :dedent: 4
     :start-after: [START howto_sensor_external_task_async]
     :end-before: [END howto_sensor_external_task_async]
+
+External Deployment Task Sensor Async
+""""""""""""""""""""""""""""""""""""""
+
+
+To check for the deployment state by making a HTTP GET request async
+:class:`~astronomer.providers.http.sensors.external_deployment_task.ExternalDeploymentTaskSensorAsync`.
+
+.. exampleinclude:: /../astronomer/providers/http/example_dags/example_external_deployment_task_sensor.py
+    :language: python
+    :dedent: 4
+    :start-after: [START howto_sensor_external_deployment_task_async]
+    :end-before: [END howto_sensor_external_deployment_task_async]
