+++ docs/providers/http/index.rst
@@ -6,8 +6,5 @@ HTTP Example DAG
   :maxdepth: 2
 
     HTTP Sensors <sensors/http>
-    External Deployment Task Sensors <sensors/external_deployment_task>
-
 
 * `HTTP Sensors <sensors/http.html>`_
-* `External Deployment Task Sensors <sensors/external_deployment_task.html>`_
