--- docs/providers/http/sensors/external_deployment_task.rst
@@ -1,12 +1,0 @@
-External Deployment Task Sensor Async
-""""""""""""""""""""""""""""""""""""""
-
-
-To check for the deployment state by making a HTTP GET request async
-:class:`~astronomer.providers.http.sensors.external_deployment_task.ExternalDeploymentTaskSensorAsync`.
-
-.. exampleinclude:: /../astronomer/providers/http/example_dags/example_external_deployment_task_sensor.py
-    :language: python
-    :dedent: 4
-    :start-after: [START howto_sensor_external_deployment_task_async]
-    :end-before: [END howto_sensor_external_deployment_task_async]
